Renaming magic link route name
